requirements.txt
# Impact Realty AI Platform - Python Dependencies
# Pin versions for production stability

# =============================================================================
# CORE FRAMEWORK
# =============================================================================
langflow==1.0.0a8
streamlit==1.28.1
fastapi==0.104.1
uvicorn[standard]==0.24.0

# =============================================================================
# DATABASE & CACHING
# =============================================================================
psycopg2-binary==2.9.7
redis==5.0.1
asyncpg==0.29.0

# =============================================================================
# HTTP & API CLIENTS
# =============================================================================
aiohttp==3.9.1
httpx==0.25.2
requests==2.31.0

# =============================================================================
# DATA VALIDATION & SERIALIZATION
# =============================================================================
pydantic==2.5.0
pydantic-settings==2.1.0

# =============================================================================
# AI & ML LIBRARIES
# =============================================================================
openai==1.3.7
langchain==0.0.341
langchain-community==0.0.1
langchain-openai==0.0.2

# =============================================================================
# AZURE SERVICES
# =============================================================================
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
azure-cosmos==4.5.1

# =============================================================================
# AUTHENTICATION & SECURITY
# =============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# =============================================================================
# UTILITIES
# =============================================================================
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30

# =============================================================================
# LOGGING & MONITORING
# =============================================================================
structlog==23.2.0
loguru==0.7.2

# =============================================================================
# TASK SCHEDULING
# =============================================================================
celery==5.3.4
croniter==2.0.1

# =============================================================================
# FILE PROCESSING
# =============================================================================
pypdf==3.17.1
python-docx==1.1.0
pillow==10.1.0
openpyxl==3.1.2

# =============================================================================
# WEB SCRAPING & DATA EXTRACTION
# =============================================================================
beautifulsoup4==4.12.2
lxml==4.9.3
scrapy==2.11.0

# =============================================================================
# TESTING
# =============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx-mock==0.10.1

# =============================================================================
# DEVELOPMENT TOOLS
# =============================================================================
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# =============================================================================
# COMMUNICATION INTEGRATIONS
# =============================================================================
twilio==8.10.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0

# =============================================================================
# SEARCH & EMBEDDINGS
# =============================================================================
tavily-python==0.3.1
sentence-transformers==2.2.2
chromadb==0.4.18
faiss-cpu==1.7.4